add the ethmsgs database table, split the eth database entity class and runtime entity class, migrate the logic, and rename the class name.
