+++ app/schemas/com.keystone.cold.db.AppDatabase/2.json
@@ -2,7 +2,7 @@
   "formatVersion": 1,
   "database": {
     "version": 2,
-    "identityHash": "c14cb8dc1c75e14a828ad8942c830129",
+    "identityHash": "57eaeb6f9dba746800b83ed8a643bd56",
     "entities": [
       {
         "tableName": "coins",
@@ -361,7 +361,7 @@
       },
       {
         "tableName": "ethtxs",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `txId` TEXT NOT NULL, `signedHex` TEXT, `from` TEXT, `timeStamp` INTEGER NOT NULL, `belongTo` TEXT, `addition` TEXT)",
+        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `txId` TEXT NOT NULL, `signedHex` TEXT, `from` TEXT, `timeStamp` INTEGER NOT NULL, `belongTo` TEXT, `txType` INTEGER NOT NULL, `addition` TEXT)",
         "fields": [
           {
             "fieldPath": "id",
@@ -400,6 +400,12 @@
             "notNull": false
           },
           {
+            "fieldPath": "txType",
+            "columnName": "txType",
+            "affinity": "INTEGER",
+            "notNull": true
+          },
+          {
             "fieldPath": "addition",
             "columnName": "addition",
             "affinity": "TEXT",
@@ -423,12 +429,59 @@
           }
         ],
         "foreignKeys": []
+      },
+      {
+        "tableName": "ethmsgs",
+        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `msgId` TEXT NOT NULL, `signature` TEXT, `timeStamp` INTEGER NOT NULL)",
+        "fields": [
+          {
+            "fieldPath": "id",
+            "columnName": "id",
+            "affinity": "INTEGER",
+            "notNull": true
+          },
+          {
+            "fieldPath": "msgId",
+            "columnName": "msgId",
+            "affinity": "TEXT",
+            "notNull": true
+          },
+          {
+            "fieldPath": "signature",
+            "columnName": "signature",
+            "affinity": "TEXT",
+            "notNull": false
+          },
+          {
+            "fieldPath": "timeStamp",
+            "columnName": "timeStamp",
+            "affinity": "INTEGER",
+            "notNull": true
+          }
+        ],
+        "primaryKey": {
+          "columnNames": [
+            "id"
+          ],
+          "autoGenerate": true
+        },
+        "indices": [
+          {
+            "name": "index_ethmsgs_id",
+            "unique": false,
+            "columnNames": [
+              "id"
+            ],
+            "createSql": "CREATE  INDEX `index_ethmsgs_id` ON `${TABLE_NAME}` (`id`)"
+          }
+        ],
+        "foreignKeys": []
       }
     ],
     "views": [],
     "setupQueries": [
       "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, 'c14cb8dc1c75e14a828ad8942c830129')"
+      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '57eaeb6f9dba746800b83ed8a643bd56')"
     ]
   }
 }
