+++ app/src/main/java/com/keystone/cold/db/entity/ETHMsgDBEntity.java
@@ -1,0 +1,50 @@
+package com.keystone.cold.db.entity;
+
+import androidx.annotation.NonNull;
+import androidx.room.Entity;
+import androidx.room.Index;
+import androidx.room.PrimaryKey;
+
+@Entity(tableName = "ethmsgs", indices = {@Index("id")})
+public class ETHMsgDBEntity {
+    @PrimaryKey(autoGenerate = true)
+    private int id;
+    @NonNull
+    private String msgId;
+    private String signature;
+    private long timeStamp;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    @NonNull
+    public String getMsgId() {
+        return msgId;
+    }
+
+    public void setMsgId(@NonNull String msgId) {
+        this.msgId = msgId;
+    }
+
+    public String getSignature() {
+        return signature;
+    }
+
+    public void setSignature(String signature) {
+        this.signature = signature;
+    }
+
+    public long getTimeStamp() {
+        return timeStamp;
+    }
+
+    public void setTimeStamp(long timeStamp) {
+        this.timeStamp = timeStamp;
+    }
+
+}
