+++ app/src/main/res/layout/eth_fee_market_tx.xml
@@ -27,7 +27,7 @@
 
         <variable
             name="tx"
-            type="com.keystone.cold.db.entity.GenericETHTxEntity" />
+            type="com.keystone.cold.viewmodel.tx.GenericETHTxEntity" />
     </data>
