+++ app/src/main/res/layout/eth_fee_market_tx_confirm.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
   ~ Copyright (c) 2021 Keystone
   ~
   ~ This program is free software: you can redistribute it and/or modify
@@ -28,7 +27,7 @@
 
         <variable
             name="tx"
-            type="com.keystone.cold.db.entity.GenericETHTxEntity" />
+            type="com.keystone.cold.viewmodel.tx.GenericETHTxEntity" />
     </data>
 
 
@@ -66,8 +65,8 @@
             <include
                 android:id="@+id/eth_tx"
                 layout="@layout/eth_fee_market_tx_detail"
-                android:layout_height="0dp"
                 android:layout_width="match_parent"
+                android:layout_height="0dp"
                 android:layout_weight="1"
                 bind:tx="@{tx}" />
 
@@ -75,8 +74,8 @@
                 android:id="@+id/sign"
                 style="@style/AcceptButton"
                 android:layout_width="match_parent"
-                android:layout_marginTop="16dp"
                 android:layout_marginHorizontal="16dp"
+                android:layout_marginTop="16dp"
                 android:layout_marginBottom="20dp"
                 android:text="@string/sign" />
         </LinearLayout>
