+++ app/src/main/res/layout/eth_fee_market_tx_detail.xml
@@ -25,7 +25,7 @@
 
         <variable
             name="tx"
-            type="com.keystone.cold.db.entity.GenericETHTxEntity" />
+            type="com.keystone.cold.viewmodel.tx.GenericETHTxEntity" />
     </data>
 
     <ScrollView
@@ -105,13 +105,13 @@
                 android:id="@+id/check_info"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
+                android:layout_marginStart="16dp"
+                android:layout_marginTop="8dp"
+                android:gravity="start"
                 android:text="@string/check_info"
                 android:textColor="@color/check_info_color"
                 android:textSize="13sp"
-                android:visibility="gone"
-                android:layout_marginStart="16dp"
-                android:layout_marginTop="8dp"
-                android:gravity="start"/>
+                android:visibility="gone" />
 
             <LinearLayout
                 android:id="@+id/tx_id_info"
@@ -144,7 +144,7 @@
                     layout="@layout/divider"
                     android:layout_width="match_parent"
                     android:layout_height="1dp"
-                    android:layout_marginTop="12sp"/>
+                    android:layout_marginTop="12sp" />
             </LinearLayout>
 
             <LinearLayout
@@ -176,7 +176,7 @@
                     layout="@layout/divider"
                     android:layout_width="match_parent"
                     android:layout_height="1dp"
-                    android:layout_marginTop="12sp"/>
+                    android:layout_marginTop="12sp" />
             </LinearLayout>
 
             <LinearLayout
@@ -209,7 +209,7 @@
                     layout="@layout/divider"
                     android:layout_width="match_parent"
                     android:layout_height="1dp"
-                    android:layout_marginTop="12sp"/>
+                    android:layout_marginTop="12sp" />
             </LinearLayout>
 
             <LinearLayout
@@ -218,19 +218,20 @@
                 android:layout_height="wrap_content"
                 android:orientation="vertical"
                 android:visibility="gone">
+
                 <include
                     android:id="@+id/ens"
+                    layout="@layout/ens_item"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="8dp"
-                    android:layout_marginTop="6dp"
-                    layout="@layout/ens_item"/>
+                    android:layout_marginTop="6dp" />
 
                 <include
                     layout="@layout/divider"
                     android:layout_width="match_parent"
                     android:layout_height="1dp"
-                    android:layout_marginTop="12sp"/>
+                    android:layout_marginTop="12sp" />
             </LinearLayout>
 
             <LinearLayout
@@ -263,7 +264,7 @@
                 layout="@layout/divider"
                 android:layout_width="match_parent"
                 android:layout_height="1dp"
-                android:layout_marginTop="12sp"/>
+                android:layout_marginTop="12sp" />
 
             <LinearLayout
                 android:id="@+id/fee_estimated"
@@ -301,11 +302,12 @@
                     android:textColor="@color/white"
                     android:textSize="12sp"
                     android:textStyle="bold" />
+
                 <include
                     layout="@layout/divider"
                     android:layout_width="match_parent"
                     android:layout_height="1dp"
-                    android:layout_marginTop="12sp"/>
+                    android:layout_marginTop="12sp" />
             </LinearLayout>
 
             <LinearLayout
@@ -350,7 +352,7 @@
                     layout="@layout/divider"
                     android:layout_width="match_parent"
                     android:layout_height="1dp"
-                    android:layout_marginTop="12sp"/>
+                    android:layout_marginTop="12sp" />
             </LinearLayout>
 
             <LinearLayout
@@ -384,8 +386,8 @@
                     android:id="@+id/info"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:orientation="horizontal"
-                    android:gravity="center_vertical">
+                    android:gravity="center_vertical"
+                    android:orientation="horizontal">
 
                     <ImageView
                         android:layout_width="15dp"
@@ -399,8 +401,8 @@
                         android:layout_marginStart="3dp"
                         android:layout_marginTop="3dp"
                         android:text="@string/undecoded_tip"
-                        android:textSize="13sp"
                         android:textColor="@color/undecoded_tip_color"
+                        android:textSize="13sp"
                         android:textStyle="bold" />
 
                     <ImageView
@@ -421,8 +423,9 @@
                 <LinearLayout
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:orientation="horizontal"
-                    android:layout_marginTop="14dp">
+                    android:layout_marginTop="14dp"
+                    android:orientation="horizontal">
+
                     <TextView
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
@@ -441,7 +444,7 @@
                         android:text="@string/from_tfcard"
                         android:textColor="@color/white"
                         android:textSize="13sp"
-                        android:visibility="gone"/>
+                        android:visibility="gone" />
 
                 </LinearLayout>
